move decorator to own file
